+++ apps/www/app/(public)/feed/page.tsx
@@ -38,7 +38,7 @@ export default async function Feed({
   return (
     <main className="pb-40">
       <section className="pt-6 w-full max-w-xl mx-auto bg-background border-muted">
-        <div className="border-y space-y-6 pt-6">
+        <div className="space-y-6">
           <HydrationBoundary state={dehydrate(queryClient)}>
             <PostList
               sort={sort}